Moved photos cli commands into own submodule
