+++ synology_cli/cli.py
@@ -1,34 +1,17 @@
 
 from click import command, echo, group, option, pass_context
+from synology_cli.cli_photos import cli_photos
 
 @group
 @pass_context
 def cli( ctx ):
     pass
 
-# main commands
+# main commands, sub commands are defined in submodules
 
-@cli.group( help='' )
-@pass_context
-def photos( ctx ):
-    pass
+cli.add_command( cli_photos )
 
 # sub commands
-
-@photos.command( 'create', help='creates photos' )
-@pass_context
-def photos_create( ctx ):
-    print( 'invoked command photos create' )
-
-@photos.command( 'list', help='lists photos' )
-@pass_context
-def photos_list( ctx ):
-    print( 'invoked command photos list' )
-
-@photos.command( 'remove', help='removes photos' )
-@pass_context
-def photos_remove( ctx ):
-    print( 'invoked command photos remove' )
 
 def main( args=None ):
     cli()  # trigger cli
